@@ -25,7 +25,7 @@
 		"jotai": "^2.0.4",
 		"js-cookie": "^3.0.1",
 		"jsonwebtoken": "^9.0.2",
-		"next": "13.1.2",
+		"next": "^14.0.3",
 		"next-auth": "^4.23.1",
 		"nodemailer": "^6.9.4",
 		"octokit": "^2.0.11",
