@@ -1,7 +1,8 @@
+"use client";
 import Button from "@components/Button";
 import Input from "@components/Input";
 import { signIn } from "next-auth/react";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 
