@@ -1,4 +1,4 @@
-import ProjectCommitService, { LatestCommitsView } from "@backend/services/ProjectCommitService";
+"use client";
 import Badge from "@components/Badge";
 import LeftHeading from "@components/Headings/LeftHeading";
 import IconLink from "@components/IconLink";
@@ -9,15 +9,7 @@ import CommitsTooltip from "@components/recharts/Tooltips/CommitsTooltip";
 import ProjectSection from "@components/Sections/ProjectSection/ProjectSection";
 import ProjectSectionBody from "@components/Sections/ProjectSection/ProjectSectionBody";
 import ProjectSectionHeading from "@components/Sections/ProjectSection/ProjectSectionHeading";
-import config from "@config/config";
 import useBreadcrumb from "@context/hooks/useBreadcrumb";
-import useCurrentUrl from "@hooks/useCurrentUrl";
-import { prisma } from "@prisma";
-import { Project } from "@prisma/client";
-import dayjs from "dayjs";
-import dayJsIsBetweenPlugin from "dayjs/plugin/isBetween";
-import { NextPageContext } from "next";
-import Head from "next/head";
 import {
 	Bar,
 	BarChart,
@@ -29,85 +21,7 @@ import {
 	XAxis,
 } from "recharts";
 
-dayjs.extend(dayJsIsBetweenPlugin);
-
-const { NEXT_PUBLIC_BASE_URL } = config;
-
-export async function getServerSideProps(context: NextPageContext) {
-	const { id } = context.query;
-
-	if (!id) {
-		return {
-			props: {
-				project: null,
-			},
-		};
-	}
-
-	const projectLanguageAggregate = await prisma.projectLanguage.aggregate({
-		where: {
-			projectId: +id,
-		},
-		_sum: {
-			codeAmountInBytes: true,
-		},
-	});
-
-	const project = await prisma.project.findFirst({
-		where: {
-			id: +id,
-		},
-		select: {
-			id: true,
-			name: true,
-			alias: true,
-			githubUrl: true,
-			imageUrl: true,
-			url: true,
-			description: true,
-			visibility: true,
-			updatedAt: true,
-			createdAt: true,
-			additionalDescription: true,
-			languages: {
-				where: {
-					codeAmountInBytes: {
-						gte: Number(
-							(projectLanguageAggregate._sum.codeAmountInBytes! * 0.01).toFixed(0)
-						),
-					},
-				},
-			},
-			commits: true,
-			tags: true,
-		},
-	});
-
-	return {
-		props: {
-			// 🤔
-			project: JSON.parse(JSON.stringify(project)) as typeof project,
-			latestCommitsView: JSON.parse(
-				JSON.stringify(
-					project
-						? ProjectCommitService.getLatestCommitsView(project)
-						: {
-								latestCommitDate: new Date(),
-								firstCommitDate: new Date(),
-								commitsGroupedByDate: [],
-						  }
-				)
-			) as LatestCommitsView,
-		} satisfies {
-			project: Project | null;
-			latestCommitsView: LatestCommitsView;
-		},
-	};
-}
-
-type ProjectDetailsProps = Awaited<ReturnType<typeof getServerSideProps>>["props"];
-
-function ProjectDetails({ project, latestCommitsView }: ProjectDetailsProps) {
+export default function ProjectDetailsPage({ project, latestCommitsView }) {
 	useBreadcrumb([
 		{
 			isHome: true,
@@ -121,9 +35,6 @@ function ProjectDetails({ project, latestCommitsView }: ProjectDetailsProps) {
 			children: project?.name || "Not found",
 		},
 	]);
-	const currentUrl = useCurrentUrl();
-
-	if (!project) return <div>no project found</div>;
 
 	const getPercentageAmountOfLanguage = (amountInBytes: number, totalAmountInBytes: number) => {
 		return Number(amountInBytes / totalAmountInBytes).toLocaleString(undefined, {
@@ -137,30 +48,8 @@ function ProjectDetails({ project, latestCommitsView }: ProjectDetailsProps) {
 		0
 	);
 
-	const metaTitle = `Project: ${project.alias}`;
-	const metaDescription = `Detailed overview of the "${project.alias}"-project`;
-	const ogImageUrl = `${NEXT_PUBLIC_BASE_URL}/api/project/${project.id}/image`;
-
 	return (
 		<>
-			<Head>
-				<title>{metaTitle}</title>
-				<meta name="description" content={metaDescription} />
-				<meta name="keywords" content={`${metaTitle}, David Kimmich`} />
-
-				{/* Facebook Meta Tags */}
-				<meta property="og:url" content={currentUrl} />
-				<meta property="og:image" content={ogImageUrl} />
-				<meta property="og:type" content="website" />
-				<meta property="og:title" content={metaTitle} />
-				<meta property="og:description" content={metaDescription} />
-				{/* Twitter Meta Tags */}
-				<meta property="twitter:url" content={currentUrl} />
-				<meta property="twitter:image" content={ogImageUrl} />
-				<meta name="twitter:title" content={metaTitle} />
-				<meta name="twitter:description" content={metaDescription} />
-			</Head>
-
 			<LeftHeading
 				rightSection={
 					<IconLink
@@ -266,5 +155,3 @@ function ProjectDetails({ project, latestCommitsView }: ProjectDetailsProps) {
 		</>
 	);
 }
-
-export default ProjectDetails;
