@@ -1,3 +1,4 @@
+import { cn } from "@utils/component-utils";
 import { ComponentPropsWithRef, PropsWithChildren } from "react";
 import { DefaultColors } from "tailwindcss/types/generated/colors";
 
@@ -17,7 +18,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "blue":
 			return (
 				<span
-					className={`rounded bg-blue-100 px-2.5 py-0.5 text-xs font-medium text-blue-800 dark:bg-blue-900 dark:text-blue-300 ${className}`}
+					className={cn(
+						`rounded bg-blue-100 px-2.5 py-0.5 text-xs font-medium text-blue-800 dark:bg-blue-900 dark:text-blue-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -26,7 +30,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "gray":
 			return (
 				<span
-					className={`rounded bg-gray-100 px-2.5 py-0.5 text-xs font-medium text-gray-800 dark:bg-gray-700 dark:text-gray-300 ${className}`}
+					className={cn(
+						`rounded bg-gray-100 px-2.5 py-0.5 text-xs font-medium text-gray-800 dark:bg-gray-700 dark:text-gray-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -35,7 +42,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "red":
 			return (
 				<span
-					className={`rounded bg-red-100 px-2.5 py-0.5 text-xs font-medium text-red-800 dark:bg-red-900 dark:text-red-300 ${className}`}
+					className={cn(
+						`rounded bg-red-100 px-2.5 py-0.5 text-xs font-medium text-red-800 dark:bg-red-900 dark:text-red-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -44,7 +54,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "green":
 			return (
 				<span
-					className={`rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300 ${className}`}
+					className={cn(
+						`rounded bg-green-100 px-2.5 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -53,7 +66,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "yellow":
 			return (
 				<span
-					className={`rounded bg-yellow-100 px-2.5 py-0.5 text-xs font-medium text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300 ${className}`}
+					className={cn(
+						`rounded bg-yellow-100 px-2.5 py-0.5 text-xs font-medium text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -62,7 +78,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "indigo":
 			return (
 				<span
-					className={`rounded bg-indigo-100 px-2.5 py-0.5 text-xs font-medium text-indigo-800 dark:bg-indigo-900 dark:text-indigo-300 ${className}`}
+					className={cn(
+						`rounded bg-indigo-100 px-2.5 py-0.5 text-xs font-medium text-indigo-800 dark:bg-indigo-900 dark:text-indigo-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -71,7 +90,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "purple":
 			return (
 				<span
-					className={`rounded bg-purple-100 px-2.5 py-0.5 text-xs font-medium text-purple-800 dark:bg-purple-900 dark:text-purple-300 ${className}`}
+					className={cn(
+						`rounded bg-purple-100 px-2.5 py-0.5 text-xs font-medium text-purple-800 dark:bg-purple-900 dark:text-purple-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
@@ -80,7 +102,10 @@ function Badge({ children, variant = "blue", className, ...props }: BadgeProps) 
 		case "pink":
 			return (
 				<span
-					className={`rounded bg-pink-100 px-2.5 py-0.5 text-xs font-medium text-pink-800 dark:bg-pink-900 dark:text-pink-300 ${className}`}
+					className={cn(
+						`rounded bg-pink-100 px-2.5 py-0.5 text-xs font-medium text-pink-800 dark:bg-pink-900 dark:text-pink-300`,
+						className
+					)}
 					{...props}
 				>
 					{children}
