@@ -1,3 +1,4 @@
+"use client";
 import loadingPortalSlotIsRenderedAtom from "@context/atoms/LoadingPortalSlotIsRenderedAtom";
 import { useAtom } from "jotai";
 import { ComponentPropsWithRef } from "react";
@@ -6,7 +7,7 @@ import LoadingCircle from "./LoadingCircle";
 
 type LoadingCircleWithPositioningProps = {} & ComponentPropsWithRef<"span">;
 
-function LoadingCircleWithPositioning({ ...props }: LoadingCircleWithPositioningProps) {
+function LoadingCircleWithPositioning(props: LoadingCircleWithPositioningProps) {
 	const [loadingPortalSlotIsRendered] = useAtom(loadingPortalSlotIsRenderedAtom);
 	return loadingPortalSlotIsRendered
 		? createPortal(<LoadingCircle />, document.getElementById("loading-portal")!)
